more unit tests
