+++ test/data/optimize/rosenbrock_mle.csv
@@ -1,0 +1,29 @@
+# stan_version_major = 2
+# stan_version_minor = 20
+# stan_version_patch = 0
+# model = rosenbrock_model
+# method = optimize
+#   optimize
+#     algorithm = lbfgs (Default)
+#       lbfgs
+#         init_alpha = 0.001 (Default)
+#         tol_obj = 9.9999999999999998e-13 (Default)
+#         tol_rel_obj = 10000 (Default)
+#         tol_grad = 1e-08 (Default)
+#         tol_rel_grad = 10000000 (Default)
+#         tol_param = 1e-08 (Default)
+#         history_size = 5 (Default)
+#     iter = 2000 (Default)
+#     save_iterations = 0 (Default)
+# id = 0 (Default)
+# data
+#   file =  (Default)
+# init = 2 (Default)
+# random
+#   seed = 1091167574
+# output
+#   file = output.csv (Default)
+#   diagnostic_file =  (Default)
+#   refresh = 100 (Default)
+lp__,x,y
+-2.80848e-10,1.00001,1.00001
