+++ test/test_model.py
@@ -4,13 +4,8 @@ import unittest
 from cmdstanpy.cmdstan_args import Method, SamplerArgs, CmdStanArgs
 from cmdstanpy.utils import EXTENSION
 from cmdstanpy.model import Model
-from cmdstanpy.stanfit import RunSet
 from contextlib import contextmanager
-import logging
-from multiprocessing import cpu_count
-import numpy as np
 import sys
-from testfixtures import LogCapture
 
 here = os.path.dirname(os.path.abspath(__file__))
 datafiles_path = os.path.join(here, 'data')
@@ -104,180 +99,6 @@ class ModelTest(unittest.TestCase):
 
     # TODO: test compile with existing exe - timestamp on exe unchanged
     # TODO: test overwrite with existing exe - timestamp on exe updated
-
-
-class SampleTest(unittest.TestCase):
-    def test_bernoulli_good(self):
-        stan = os.path.join(datafiles_path, 'bernoulli.stan')
-        exe = os.path.join(datafiles_path, 'bernoulli' + EXTENSION)
-        bern_model = Model(stan_file=stan, exe_file=exe)
-        bern_model.compile()
-
-        jdata = os.path.join(datafiles_path, 'bernoulli.data.json')
-        bern_fit = bern_model.sample(
-            data=jdata, chains=4, cores=2, seed=12345, sampling_iters=100
-        )
-        self.assertEqual(bern_fit.runset._args.method, Method.SAMPLE)
-
-        for i in range(bern_fit.runset.chains):
-            csv_file = bern_fit.runset.csv_files[i]
-            txt_file = ''.join([os.path.splitext(csv_file)[0], '.txt'])
-            self.assertTrue(os.path.exists(csv_file))
-            self.assertTrue(os.path.exists(txt_file))
-
-        self.assertEqual(bern_fit.runset.chains, 4)
-        self.assertEqual(bern_fit.draws, 100)
-        column_names = [
-            'lp__',
-            'accept_stat__',
-            'stepsize__',
-            'treedepth__',
-            'n_leapfrog__',
-            'divergent__',
-            'energy__',
-            'theta',
-        ]
-        self.assertEqual(bern_fit.column_names, tuple(column_names))
-
-        bern_sample = bern_fit.sample
-        self.assertEqual(bern_sample.shape, (100, 4, len(column_names)))
-
-        self.assertEqual(bern_fit.metric_type, 'diag_e')
-        self.assertEqual(bern_fit.stepsize.shape, (4,))
-        self.assertEqual(bern_fit.metric.shape, (4, 1))
-
-        output = os.path.join(datafiles_path, 'test1-bernoulli-output')
-        bern_fit = bern_model.sample(
-            data=jdata,
-            chains=4,
-            cores=2,
-            seed=12345,
-            sampling_iters=100,
-            csv_basename=output,
-        )
-        for i in range(bern_fit.runset.chains):
-            csv_file = bern_fit.runset.csv_files[i]
-            txt_file = ''.join([os.path.splitext(csv_file)[0], '.txt'])
-            self.assertTrue(os.path.exists(csv_file))
-            self.assertTrue(os.path.exists(txt_file))
-        bern_sample = bern_fit.sample
-        self.assertEqual(bern_sample.shape, (100, 4, len(column_names)))
-        for i in range(bern_fit.runset.chains):  # cleanup datafile_path dir
-            os.remove(bern_fit.runset.csv_files[i])
-            os.remove(bern_fit.runset.console_files[i])
-
-        rdata = os.path.join(datafiles_path, 'bernoulli.data.R')
-        bern_fit = bern_model.sample(
-            data=rdata, chains=4, cores=2, seed=12345, sampling_iters=100
-        )
-        bern_sample = bern_fit.sample
-        self.assertEqual(bern_sample.shape, (100, 4, len(column_names)))
-
-        data_dict = {'N': 10, 'y': [0, 1, 0, 0, 0, 0, 0, 0, 0, 1]}
-        bern_fit = bern_model.sample(
-            data=data_dict, chains=4, cores=2, seed=12345, sampling_iters=100
-        )
-        bern_sample = bern_fit.sample
-        self.assertEqual(bern_sample.shape, (100, 4, len(column_names)))
-
-    def test_bernoulli_bad(self):
-        stan = os.path.join(datafiles_path, 'bernoulli.stan')
-        exe = os.path.join(datafiles_path, 'bernoulli' + EXTENSION)
-        bern_model = Model(stan_file=stan, exe_file=exe)
-        bern_model.compile()
-
-        with self.assertRaisesRegex(Exception, 'Error during sampling'):
-            bern_fit = bern_model.sample(
-                chains=4, cores=2, seed=12345, sampling_iters=100
-            )
-
-    def test_multi_proc(self):
-        logistic_stan = os.path.join(datafiles_path, 'logistic.stan')
-        logistic_model = Model(stan_file=logistic_stan)
-        logistic_model.compile()
-        logistic_data = os.path.join(datafiles_path, 'logistic.data.R')
-
-        with LogCapture() as log:
-            logger = logging.getLogger()
-            fit = logistic_model.sample(data=logistic_data, chains=4, cores=1)
-        log.check_present(
-            ('cmdstanpy', 'INFO', 'finish chain 1'),
-            ('cmdstanpy', 'INFO', 'start chain 2'),
-        )
-        with LogCapture() as log:
-            logger = logging.getLogger()
-            fit = logistic_model.sample(data=logistic_data, chains=4, cores=2)
-        if cpu_count() >= 4:
-            # finish chains 1, 2 before starting chains 3, 4
-            log.check_present(
-                ('cmdstanpy', 'INFO', 'finish chain 1'),
-                ('cmdstanpy', 'INFO', 'start chain 4'),
-            )
-        if cpu_count() >= 4:
-            with LogCapture() as log:
-                logger = logging.getLogger()
-                fit = logistic_model.sample(
-                    data=logistic_data, chains=4, cores=4
-                )
-                log.check_present(
-                    ('cmdstanpy', 'INFO', 'start chain 4'),
-                    ('cmdstanpy', 'INFO', 'finish chain 1'),
-                )
-
-
-class GenerateQuantitiesTest(unittest.TestCase):
-    def test_gen_quantities_good(self):
-        stan = os.path.join(datafiles_path, 'bernoulli_ppc.stan')
-        model = Model(stan_file=stan)
-        model.compile()
-
-        jdata = os.path.join(datafiles_path, 'bernoulli.data.json')
-
-        # synthesize runset for existing sample
-        goodfiles_path = os.path.join(datafiles_path, 'runset-good')
-        output = os.path.join(goodfiles_path, 'bern')
-        sampler_args = SamplerArgs(
-            sampling_iters=100, max_treedepth=11, adapt_delta=0.95
-        )
-        cmdstan_args = CmdStanArgs(
-            model_name=model.name,
-            model_exe=model.exe_file,
-            chain_ids=[1, 2, 3, 4],
-            seed=12345,
-            data=jdata,
-            output_basename=output,
-            method_args=sampler_args,
-        )
-        sampler_runset = RunSet(args=cmdstan_args, chains=4)
-        for i in range(4):
-            sampler_runset._set_retcode(i, 0)
-
-        bern_gqs = model.run_generated_quantities(
-            csv_files=sampler_runset.csv_files, data=jdata
-        )
-        self.assertEqual(
-            bern_gqs.runset._args.method, Method.GENERATE_QUANTITIES
-        )
-
-        # check results - ouput files, quantities of interest, draws
-        self.assertEqual(bern_gqs.runset.chains, 4)
-        for i in range(bern_gqs.runset.chains):
-            self.assertEqual(bern_gqs.runset._retcode(i), 0)
-            csv_file = bern_gqs.runset.csv_files[i]
-            self.assertTrue(os.path.exists(csv_file))
-        column_names = [
-            'y_rep.1',
-            'y_rep.2',
-            'y_rep.3',
-            'y_rep.4',
-            'y_rep.5',
-            'y_rep.6',
-            'y_rep.7',
-            'y_rep.8',
-            'y_rep.9',
-            'y_rep.10',
-        ]
-        self.assertEqual(bern_gqs.column_names, tuple(column_names))
 
 
 if __name__ == '__main__':
